add areaInfoSheet component, rename and move component addNewProjectcard
